feat: add remov img from bookmarks
