+++ client/src/components/Main/Bookmarks/Bookmarks.scss
@@ -1,0 +1,3 @@
+.img {
+  object-fit: cover;
+}
