+++ client/src/components/Main/Found/Found.scss
@@ -1,0 +1,3 @@
+.img {
+  object-fit: cover;
+}
